fix: addressed all requested changes
